# frontend/openapi/openapi.snapshot.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/client-contracts": {
      "get": {
        "operationId": "ClientContractController_findAll",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "example": 10,
              "type": "number"
            }
          },
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "sortBy",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "contractNumber",
                "startDate",
                "createdAt",
                "updatedAt"
              ],
              "type": "string"
            }
          },
          {
            "name": "sortOrder",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "asc",
                "desc"
              ],
              "type": "string"
            }
          },
          {
            "name": "clientId",
            "required": false,
            "in": "query",
            "schema": {
              "format": "uuid"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "ACTIVE",
                "EXPIRED",
                "CANCELLED"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated list. Actual response wraps payload in `{ success, timestamp, data: { data, meta } }`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClientContractListEnvelopeOpenApiDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid query",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardErrorResponseDto"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "List client contracts (pagination)",
        "tags": [
          "Client contracts"
        ]
      },
      "post": {
        "description": "Requires a valid JWT. No persisted link to User — only the authenticated caller is implied by the request.",
        "operationId": "ClientContractController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateClientContractBodyDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Contract created. Actual response wraps payload in `{ success, timestamp, data }`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClientContractSingleEnvelopeOpenApiDto"
                }
              }
            }
          },
          "400": {
            "description": "Zod validation / domain rules",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Client not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardErrorResponseDto"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Create client contract",
        "tags": [
          "Client contracts"
        ]
      }
    },
    "/api/v1/client-contracts/{id}": {
      "get": {
        "operationId": "ClientContractController_findById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Contract found. Actual response wraps payload in `{ success, timestamp, data }`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClientContractSingleEnvelopeOpenApiDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Contract not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardErrorResponseDto"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Get client contract by id",
        "tags": [
          "Client contracts"
        ]
      },
      "patch": {
        "operationId": "ClientContractController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateClientContractBodyDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Contract updated. Actual response wraps payload in `{ success, timestamp, data }`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClientContractSingleEnvelopeOpenApiDto"
                }
              }
            }
          },
          "400": {
            "description": "Zod validation / domain rules",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Contract not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardErrorResponseDto"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Update client contract",
        "tags": [
          "Client contracts"
        ]
      }
    }
  },
  "info": {
    "title": "DL System API",
    "description": "DL System HTTP API: JWT access tokens and httpOnly refresh cookies. Responses use `{ success, timestamp, data }` (or error envelope). **Tickets**, **clients**, and **client contracts** require `Authorization: Bearer <accessToken>`. Register at `POST /users`, then `POST /auth/login`.",
    "version": "0.0.1",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "/api/v1",
      "description": "Versioned API"
    }
  ],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "JWT from `POST /auth/login` (`data.accessToken`). Payload: `sub` (user uuid), `email`. Refresh: httpOnly cookie + `POST /auth/refresh`."
      }
    },
    "schemas": {
...
      "StandardErrorResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "timestamp": {
            "type": "string",
            "example": "2025-03-25T12:00:00.000Z"
          },
          "statusCode": {
            "type": "number",
            "example": 400
          },
          "error": {
            "type": "string",
            "example": "Bad Request"
          },
          "message": {
            "oneOf": [
              {
                "type": "string"
              },
              {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            ]
          },
          "details": {
            "type": "object"
          }
        },
        "required": [
          "success",
          "timestamp",
          "statusCode",
          "error",
          "message"
        ]
      },
...
      "PaginationMetaOpenApiDto": {
        "type": "object",
        "properties": {
          "total": {
            "type": "number"
          },
          "page": {
            "type": "number"
          },
          "limit": {
            "type": "number"
          },
          "totalPages": {
            "type": "number"
          },
          "hasNextPage": {
            "type": "boolean"
          },
          "hasPreviousPage": {
            "type": "boolean"
          },
          "nextCursor": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "total",
          "page",
          "limit",
          "totalPages",
          "hasNextPage",
          "hasPreviousPage",
          "nextCursor"
        ]
      },
...
      "ClientAddressOpenApiDto": {
        "type": "object",
        "properties": {
          "street": {
            "type": "string"
          },
          "number": {
            "type": "string"
          },
          "complement": {
            "type": "string"
          },
          "neighborhood": {
            "type": "string"
          },
          "city": {
            "type": "string"
          },
          "state": {
            "type": "string",
            "minLength": 2,
            "maxLength": 2
          },
          "zipCode": {
            "type": "string",
            "description": "8 digits"
          }
        },
        "required": [
          "street",
          "number",
          "neighborhood",
          "city",
          "state",
          "zipCode"
        ]
      },
...
      "ClientContractPublicHttpOpenApiDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "contractNumber": {
            "type": "string"
          },
          "clientId": {
            "type": "string",
            "format": "uuid"
          },
          "useClientAddress": {
            "type": "boolean"
          },
          "address": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/ClientAddressOpenApiDto"
              }
            ]
          },
          "startDate": {
            "type": "string",
            "format": "date"
          },
          "endDate": {
            "type": "object",
            "format": "date",
            "nullable": true
          },
          "status": {
            "type": "string",
            "enum": [
              "ACTIVE",
              "EXPIRED",
              "CANCELLED"
            ]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "contractNumber",
          "clientId",
          "useClientAddress",
          "address",
          "startDate",
          "endDate",
          "status",
          "createdAt",
          "updatedAt"
        ]
      },
      "ClientContractListInnerOpenApiDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ClientContractPublicHttpOpenApiDto"
            }
          },
          "meta": {
            "$ref": "#/components/schemas/PaginationMetaOpenApiDto"
          }
        },
        "required": [
          "data",
          "meta"
        ]
      },
      "ClientContractListEnvelopeOpenApiDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "timestamp": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/ClientContractListInnerOpenApiDto"
          }
        },
        "required": [
          "success",
          "timestamp",
          "data"
        ]
      },
      "ClientContractSingleEnvelopeOpenApiDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "timestamp": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/ClientContractPublicHttpOpenApiDto"
          }
        },
        "required": [
          "success",
          "timestamp",
          "data"
        ]
      },
      "CreateClientContractBodyDto": {
        "type": "object",
        "properties": {
          "contractNumber": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100
          },
          "clientId": {
            "type": "string",
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
          },
          "useClientAddress": {
            "default": true,
            "type": "boolean"
          },
          "address": {
            "type": "object",
            "properties": {
              "street": {
                "type": "string",
                "minLength": 1
              },
              "number": {
                "type": "string",
                "minLength": 1
              },
              "complement": {
                "type": "string"
              },
              "neighborhood": {
                "type": "string",
                "minLength": 1
              },
              "city": {
                "type": "string",
                "minLength": 1
              },
              "state": {
                "type": "string",
                "minLength": 2,
                "maxLength": 2
              },
              "zipCode": {
                "type": "string",
                "minLength": 1
              }
            },
            "required": [
              "street",
              "number",
              "neighborhood",
              "city",
              "state",
              "zipCode"
            ],
            "x-nestjs_zod-self-required": false
          },
          "startDate": {
            "type": "string",
            "pattern": "^\\d{4}-\\d{2}-\\d{2}$"
          },
          "endDate": {
            "type": "string",
            "pattern": "^\\d{4}-\\d{2}-\\d{2}$"
          }
        },
        "required": [
          "contractNumber",
          "clientId",
          "startDate"
        ]
      },
      "UpdateClientContractBodyDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "ACTIVE",
              "EXPIRED",
              "CANCELLED"
            ],
            "x-nestjs_zod-has-null": true
          },
          "endDate": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^\\d{4}-\\d{2}-\\d{2}$"
              },
              {
                "type": "null"
              }
            ],
            "x-nestjs_zod-has-null": true,
            "x-nestjs_zod-empty-type": true
          },
          "useClientAddress": {
            "type": "boolean",
            "x-nestjs_zod-has-null": true
          },
          "address": {
            "type": "object",
            "properties": {
              "street": {
                "type": "string",
                "minLength": 1
              },
              "number": {
                "type": "string",
                "minLength": 1
              },
              "complement": {
                "type": "string"
              },
              "neighborhood": {
                "type": "string",
                "minLength": 1
              },
              "city": {
                "type": "string",
                "minLength": 1
              },
              "state": {
                "type": "string",
                "minLength": 2,
                "maxLength": 2
              },
              "zipCode": {
                "type": "string",
                "minLength": 1
              }
            },
            "required": [
              "street",
              "number",
              "neighborhood",
              "city",
              "state",
              "zipCode"
            ],
            "x-nestjs_zod-has-null": true,
            "x-nestjs_zod-self-required": false
          }
        }
      }
    }
  }
}